API

Contents

API

Contents

Index

Templates

Images

Movies

Divergences

Extractor

Utilities

Visualizations

Index

VIDA.AbstractCompositeTemplate

VIDA.AbstractDivergence

VIDA.AbstractFitsImage

VIDA.AbstractImage

VIDA.AbstractImageTemplate

VIDA.AbstractModifierTemplate

VIDA.AbstractMovie

VIDA.AbstractTemplate

VIDA.AddTemplate

VIDA.AsymGaussian

VIDA.BBO

VIDA.Bhattacharyya

VIDA.CMAES

VIDA.Constant

VIDA.CosineRing

VIDA.Disk

VIDA.EHTImage

VIDA.EHTMovie

VIDA.EllipticalGaussianRing

VIDA.ExtractProblem

VIDA.GaussianRing

VIDA.GeneralGaussianRing

VIDA.ImageTemplate

VIDA.KullbackLeibler

VIDA.LogSpiral

VIDA.MulTemplate

VIDA.Opt

VIDA.Optimizer

VIDA.RotateMod

VIDA.SlashedGaussianRing

VIDA.StretchMod

VIDA.SymCosineRing

VIDA.TIDAGaussianRing

Base.size

Base.split

VIDA.blur

VIDA.blur

VIDA.centroid

VIDA.clipimage

VIDA.extractor

VIDA.field_of_view

VIDA.flux

VIDA.flux

VIDA.get_frames

VIDA.get_image

VIDA.get_times

VIDA.inertia

VIDA.join_frames

VIDA.load_fits

VIDA.load_hdf5

VIDA.make_image

VIDA.pixelloc

VIDA.rescale

VIDA.rescale

VIDA.rotate

VIDA.save_fits

VIDA.save_hdf5

VIDA.stack

VIDA.stretch

VIDA.stretchrotate

VIDA.template_image

VIDA.threaded_extractor

VIDA.triptic

VIDA.unpack

Templates

These are the various template types and helper functions.

VIDA.AbstractTemplate — Type

abstract type AbstractTemplate

An abstract type that defines super template type.

VIDA.AbstractImageTemplate — Type

abstract type AbstractImageTemplate <: VIDA.AbstractTemplate

An abstract type that will contain the template information, such as the parameters. Specific instanstantiations will need to be defined for you to use this.

Details

This defined the highest function type. If you wish to implement your own template you
need to define a a couple of things
1. The template type <: AbstractTemplate
2. an functor of the type that computes the template function
3. an `Base.size` function that defines the number of parameters of the template.

An example is given by:

#All of our composite type are defined using the Paramters.jl package to you
can directly refer to the struct parameters when creating it, although this isn't
actually used anywhere in the code.
@with_kw struct Gaussian <: AbstractImageTemplate
    σ::Float64
    x0::Float64
    y0::Float64
end

#Typically we inline and force the function to use fastmath
 @inline function (θ::Gaussian)(x,y)
    return 1.0/(2π*σ^2)*exp(-0.5*( (x-x0)^2+(y-y0)^2 ) )
end
Base.size(::Type{Gaussian}) = 3

VIDA.AbstractModifierTemplate — Type

abstract type AbstractModifierTemplate <: VIDA.AbstractTemplate

Meta-template that accepts a abstract template and modifies it through some analytic transformation. This is useful for model composability.

VIDA.AbstractCompositeTemplate — Type

abstract type AbstractCompositeTemplate <: VIDA.AbstractTemplate

Abstract type for taking two templates composing them in some manner and returning a new template.

VIDA.ImageTemplate — Type

Type for an image template. This takes an EHTImage or any image object and creates a template out of it. The parameters of the image are the center of the image x0, y0.

Fields

x0

y0

itp

VIDA.LogSpiral — Type

Template type for a logarithmic spiral segment

Fields

r0

Radius of the spiral peak brightness

κ

Unit curvature of the logarithmic spiral

σ

thickness of the Gaussian spiral arm

δϕ

Azimuthal extent of the spiral arm

ξ

peak brightness location

x0

x location of disk center in μas

y0

y location of disk center in μas

VIDA.Constant — Type

struct Constant <: VIDA.AbstractImageTemplate

An constant template.

Details

Defines an image that just has constant flux. This is very useful for soaking up low levels of flux in image reconstructions that can bias the results.

Since images or normalized to unity, this means the Constant template has no additional parameters.

VIDA.Disk — Type

struct Disk <: VIDA.AbstractImageTemplate

A smoothed disk model

Details

Defines a template for an image that has a smoothed disk model.

VIDA.AsymGaussian — Type

struct AsymGaussian <: VIDA.AbstractImageTemplate

An asymmetric Gaussian blob.

Details

Defines a asymmetric Gaussian image. This is useful if the image has some non-ring emission in the and you need to soak up some of the flux.

The parameters of the model follow very closely to those used in Themis. The Gaussian size σ is given by σ = √(σxσy) τ = 1-σy/σx, where σx,σy are the semi-major,minor axis lenght respectively. This is similar to how the asymmetry for the EllipticalGaussianRing.

Fields

σ

Gaussian size in μas

τ

Gaussian asymmetry

ξ

Gaussian orientation in radians measured north of east

x0

x position of Gaussian center in μas

y0

y position of Gaussian center in μas

VIDA.GaussianRing — Type

struct GaussianRing <: VIDA.AbstractImageTemplate

Symmetric gaussian ring template. This is the most basic template and just attempts to recover a location x0,y0, radius r0 and thickness σ from some image.

Fields

r0

Radius of Gaussian ring in μas

σ

Standard deviation of Gaussian ring in μas

x0

x location of the center of the Gaussian ring in μas

y0

y location of the center of the Gaussian ring in μas

Example

GaussianRing(r0=20.0,σ=5.0,x0=0.0,y0=-10.0)

VIDA.SlashedGaussianRing — Type

struct SlashedGaussianRing <: VIDA.AbstractImageTemplate

Implements the slashed gaussian ring template, that uses a cosine to symmetrically implement the slash. While this is marginally more complicated that a linear slash, it has a number of benefits such as mainting the azimuthal and smooth structure of the image.

Fields

r0

Radius of the ring in μas

σ

Standard deviation of Gaussian ring in μas

s

Slash strength of Gaussiang ring. 0 means no slash

ξ

Rotation angle in radians of slash direction, measured north of west

x0

x position of the center of the ring in μas

y0

y position of the center of the ring in μas

VIDA.EllipticalGaussianRing — Type

struct EllipticalGaussianRing <: VIDA.AbstractImageTemplate

Implements the elliptical gaussian ring template. Where the ellipticity tau is defined as one minus ratio between the semi-minor and semi-major axis.

Details

Adds ellipticity to the ring. The radius r0 of the ring is now defined as the geometric mean of the semi-major (a) and semi-minor (b) axis lengths i.e.

r0 = √(a*b).

The ellipticity τ is given by τ = 1-b/a.

Fields

r0

Radius of the Gaussian ring

σ

Standard deviation of the width of the Gaussian ring

τ

Asymmetry of the Gaussian ring defined as $1-b/a$

ξ

Asymmetry orientation in radians measured north of east

x0

x position of the center of the ring in μas

y0

y position of the center of the ring in μas

Notes

There is no normalization since the ellipticity makes it impossible to normalize analytically. In fact the distance from the ellipse is implemented numerically using an algorithm adapted from git

VIDA.TIDAGaussianRing — Type

struct TIDAGaussianRing <: VIDA.AbstractImageTemplate

Creates the template from the Paper I am writing. It is a combination of the elliptical and slashed gaussian ring. The slash and the semi-major axis are either aligned if the slash parameter s>0 or antialigned if s<0.

Details

Adds ellipticity to the ring. The radius r0 of the ring is now defined as the geometric mean of the semi-major (a) and semi-minor (b) axis lengths i.e.

r0 = √(a*b).

The ellipticity τ is given by τ = 1-b/a.

Fields

r0

Radius of the Gaussian ring

σ

Standard deviation of the width of the Gaussian ring

τ

Asymmetry of the Gaussian ring defined as $1-b/a$

s

Slash of Gaussian ring.

ξ

Slash/Asymmetry orientation in radians measured north of east

x0

x position of the center of the ring in μas

y0

y position of the center of the ring in μas

VIDA.GeneralGaussianRing — Type

struct GeneralGaussianRing <: VIDA.AbstractImageTemplate

Creates the most general elliptical slashed gaussian ring model. It is a combination of the elliptical and slashed gaussian ring. The direction of the slash and the ellipticity are not aligned or anti-aligned like with the TIDAGaussianRing type.

Details

Adds ellipticity to the ring. The radius r0 of the ring is now defined as the geometric mean of the semi-major (a) and semi-minor (b) axis lengths i.e.

r0 = √(a*b).

The ellipticity τ is given by τ = 1-b/a.

Fields

r0

Radius of the Gaussian ring

σ

Standard deviation of the width of the Gaussian ring

τ

Asymmetry of the Gaussian ring defined as $1-b/a$

ξτ

Asymmetry orientation in radians, measured north of east

s

Slash of Gaussian ring.

ξs

Slash orientation in radians measured north of east

x0

x position of the center of the ring in μas

y0

y position of the center of the ring in μas

VIDA.SymCosineRing — Type

struct SymCosineRing{N, M} <: VIDA.AbstractImageTemplate

Extrememly flexible symmetric ring model. The thickness is modeled as a cosine expansion with N terms and the slash by a expansion with M terms.

Details

The ring is forced to be symmetric for a significant speed boost over CosineRing. The thickness of the ring is modeled by a cosine expansion in azimuthal angle. N specifies the number of cosine modes to fit, where the first mode is the constant thickness portion and so has no corresponding angle. The slash is modeled as a separate cosine expansion, with M terms. Here the zero order term is forced to be unity, so M defines the M additional terms.

VIDA.CosineRing — Type

struct CosineRing{N, M} <: VIDA.AbstractImageTemplate

Extrememly flexible ring model. The thickness is modeled as a cosine expansion with N terms and the slash by a expansion with M terms.

Details

The ring is allowed to be elliptical. The thickness of the ring is modeled by a cosine expansion in azimuthal angle. N specifies the number of cosine modes to fit, where the first mode is the constant thickness portion and so has no corresponding angle. The slash is modeled as a separate cosine expansion, with M terms. Here the zero order term is forced to be unity, so M defines the M additional terms.

VIDA.AddTemplate — Type

struct AddTemplate{T1<:VIDA.AbstractTemplate, T2<:VIDA.AbstractTemplate} <: VIDA.AbstractCompositeTemplate

Combines two templates together into one object. Since addition is assoiciative this can actually we used to hold multiple different templates.

Details

Overloads the Base.:+ function so you can easily add two templates together.

Example

θ1 = GaussianRing(10,5,0,0)
θ2 = SlashedGaussianRing(15,5,0.5,π/4,0,0)
θ12 = θ1+θ2

VIDA.MulTemplate — Type

struct MulTemplate{T<:VIDA.AbstractTemplate, S<:Number} <: VIDA.AbstractModifierTemplate

Multiplies template by a constant. This is useful when combining with AddTemplate since it will change the relative weights of each template.

Details

Overloads the Base.:* function so you can easily multiple a template by a number.

Example

θ = GaussianRing(15,5,0.0,0.0)
2*θ

VIDA.StretchMod — Type

StretchMod(θ::AbstractTemplate, τ)

Modifies a filter by stretching it in the x-direction by 1/√(1-τ) and y-direction by sqrt(1-τ). Typically call the stretch function instead.

VIDA.stretch — Function

stretch(m, τ)

Stretch the template m, using the τ factor. This will stretch the x-axis by 1/√(1-τ) and y-axis by √(1-τ).

VIDA.RotateMod — Type

RotateMod(θ::AbstractTemplate, ξ)

Modifies a filter by rotating it using the rotation matrix with angle ξ. Typically do not call this function directly but instread the rotate function.

VIDA.rotate — Function

rotate(m, ξ)

Rotate the template m by ξ. This rotates north of east.

VIDA.stretchrotate — Function

stretchrotate(m, τ, ξ)

Stretch and then rotate a template producing a modified template. This is equivalent to rotate(stretch(m, τ), ξ).

Base.size — Method

size(f::AbstractTemplate)

Get the number of parameters for the template f

VIDA.unpack — Function

unpack(θinit)

Unpacks the parameters of the template θ

Returns the parameters in a vector.

VIDA.stack — Function

stack(θ, θ1)

Stacks templates together so you can easily combine multiple templates. It does this by calling the :+ and :* method. Every template added will include an additional parameter that controls the relative weight of each template.

Base.split — Function

split(θ)

Splits the template into an array with its subcomponents so you can easily access them.

VIDA.template_image — Function

template_image(θ, npix, xlim, ylim)

Creates an npix×npix rasterized image of the template θ with limits xlim and ylim

Returns the tuple (xitr,yitr,image) where xitr,yitr are the iterators defining the pixel locations (which are centered) and the rasterized image, in Jy/μas^2.

Note

I use the pixel size definition fieldofview/npix, but the image is evaluated at the pixel centers.

We also use the astronomer orientation and ordering.

Images

VIDA has an image interface that reads in images using the FITS standard.

VIDA.AbstractImage — Type

An abstact type that acts as a wrapper for image objects used in astronomy.

This is the top of the castle for images and will be rarely touched. Basically unless you don't want to use fits images this will not be used

VIDA.AbstractFitsImage — Type

An absract image that will hold a fits image after being created or parsed in. This will form the basis for most astronomical images that are defined.

The defines an interface to the FITS image methods below. In the future I may change this to use the traits interfaces.

VIDA.EHTImage — Type

struct EHTImage{F, T<:AbstractArray{F, 2}} <: VIDA.AbstractFitsImage{F, T<:AbstractArray{F, 2}}

Details

The trait is to hold a EHT image tyically in matrix form. Namely the trait will typically be Matrix{Float64}.

nx is the number of pixels in the x or RA direction ny is the number of pixels in the y or DEC direction psize_x, psize_y are the pixel sizes in the x and y direction source is the source we are looking at e.g. M87 ra,dec are the sources RA and DEC in J2000 coordinates using degrees wavelength is the wavelength of the image. mjd is the Modified Julian Date of the observation. img is the actual pixeled image in Jy/pixel

VIDA.load_fits — Function

load_fits(fits_name)

where fits_name should be a fits file generated using ehtim

Details

This reads in a fits file that is more robust to the various imaging algorithms in the EHT, i.e. is works with clean, smili, eht-imaging.

The function returns an EHTImage object that contains the relevant image and parameters extracted from the fits file. It also ensures that we are astronomers and that the image using sky-left coordinates.

VIDA.save_fits — Function

save_fits(image::EHTImage, fname::String)

Save the image as a fits object with filename fname

VIDA.centroid — Function

centroid(img)

Finds the centroid or center of light of the img in μas.

VIDA.inertia — Function

inertia(img)
inertia(img, center)

Find the image moment of inertia or second moment

Notes

If center=true then we find the central second moment, or the second cumulant of the image.

VIDA.pixelloc — Function

pixelloc(img)

Returns two iterators (ra,dec) that give the locations of the img pixels.

VIDA.flux — Method

flux(img)

Finds the image flux of an EHTImage img

VIDA.rescale — Method

rescale(img::EHTImage, npix, xlim, ylim)

Inputs

img::EHTImage : Image you want to rescale

npix : Number of pixels in x and y direction

xlim : Tuple with the limits of the image in the RA in μas

ylim : Tuple with the limits of the image in DEC in μas

VIDA.blur — Method

blur(img, fwhm)

Blurs the img with a gaussian kernel with fwhm in μas. If fwhm is a scalar then the kernel is assumed to be symmetric, otherwise you the first entry is the fwhm in the EW direction and second the NS direction.

Returns the blurred image.

Missing docstring.

Missing docstring for Base.size(img::EHTImage). Check Documenter's build log for details.

VIDA.clipimage — Function

clipimage(clip, image)
clipimage(clip, image, mode)

Clips the image im according to the value clip. There are two modes for image clipping: - :relative which zeros the pixels whose intensity are below clip relative to the max. - :absolute which zeros the pixels whose intensity is below clip in Jy/pixel

VIDA.field_of_view — Function

field_of_view(img)

Finds the field of view of an EHTImage. Return a w element tuple with the field of view in the x and y direction

Movies

VIDA also has a movie interface using hdf5. Note that movies are more than just a list of images. We also use an interpolation between frames.

VIDA.AbstractMovie — Type

abstract type AbstractMovie

Creates an abstract movie class

VIDA.EHTMovie — Type

struct EHTMovie{F, T<:(Interpolations.AbstractExtrapolation{F, N, ITPT, IT} where {N, ITPT, IT})} <: VIDA.AbstractMovie

Details

The type is to hold a EHT movie. The dimension of the movie array is assumed to be in the form DEC,RA,Time.

nx is the number of pixels in the x or RA direction ny is the number of pixels in the y or DEC direction psize_x, psize_y are the pixel sizes in the x and y direction source is the source we are looking at e.g. M87 ra,dec are the sources RA and DEC in J2000 coordinates using degrees wavelength is the wavelength of the image. mjd is the Modified Julian Date of the observation. frames is the interpolation object that hold the movie frames

VIDA.load_hdf5 — Function

load_hdf5(filename; style)

Loads an hdf5 file where filename should be a HDF5 file.

Details

This reads in a hdf5 file and outputs and EHTMovie object.

Notes

Currently this only works with movies created by ehtim. SMILI uses a different format, as does Illinois, and every other group...

VIDA.save_hdf5 — Function

save_hdf5(filename, mov; style)

Saves and hdf5 file where filename is the write out location. Currently style only works with ehtim, namely we save HDF5 files that only work with ehtim.

VIDA.join_frames — Function

join_frames(times, images)

Joins an array of EHTImages at specified times to form an EHTMovie object.

Inputs

times: An array of times that the image was created at

images: An array of EHTImage objects

Outputs

EHTMovie object

VIDA.get_times — Function

get_times(mov)

Returns the times that the movie object mov was created at. This does not have to be uniform in time.

VIDA.get_frames — Function

get_frames(mov)

Gets all the frames of the movie object mov. This returns a array of EHTImage objects.

VIDA.get_image — Function

get_image(mov, t)

Gets the frame of the movie object mov at the time t. This returns an EHTImage object at the requested time. The returned object is found by linear interpolation.

VIDA.flux — Method

flux(mov, t)

Returns the flux of the mov at the times time in fractional hours

VIDA.blur — Method

blur(mov, fwhm)

Blurs the mov with a gaussian kernel with fwhm in μas. If fwhm is a scalar then the kernel is assumed to be symmetric, otherwise you the first entry is the fwhm in the EW direction and second the NS direction.

Returns the blurred movie.

VIDA.rescale — Method

rescale(mov::EHTMovie, npix, xlim, ylim)

Inputs

mov::EHTMovie : Movie you want to rescale

npix : Number of pixels in x and y direction

xlim : Tuple with the limits of the image in the RA

ylim : Tuple with the limits of the image in DEC

Divergences

VIDA.AbstractDivergence — Type

abstract type AbstractDivergence

An abstract class for a divergence of a function. This expects that a subtype has a field with an EHTImage object and a flux type. The struct is then assumed to be a functor and have a function that computes the divergence of the image and a template.

For example

    struct MyDiv{T,F,S} <: AbstractDivergence
        img::EHTImage{T,F}
        flux::S
    end
    function (bh::MyDiv)(θ::AbstractTemplate)
        ...
    end

VIDA.Bhattacharyya — Type

struct Bhattacharyya{T<:EHTImage, S} <: VIDA.AbstractDivergence

Type for the Bhattacharyya divergence. It constructed from an EHTImage i.e. data. Additionally to evaluate the divergence we use a functor approach where if θ is your

Details

This computes the Bhattacharyya divergence which is related to Hellinger distance between two distributions. In fact, they are both minimized at the same point. The Bhattacharyya divergence is defined as

\[Bh(f_\theta||\hat{I}) = -\log\int \sqrt{f_\theta(x,y)\hat{I}(x,y)}dxdy,\]

where $\hat{I}$ is defined as the image normalized to unit flux.

VIDA.KullbackLeibler — Type

struct KullbackLeibler{T, S} <: VIDA.AbstractDivergence

Type for the KL divergence. It constructed from an EHTImage i.e. data. Additionally to evaluate the divergence we use a functor approach where if θ is your

Details

This computes the KL divergence which is related to Hellinger distance between two distributions. In fact, they are both minimized at the same point. The Bhattacharyya divergence is defined as

\[KL(f_\theta||\hat{I}) = -\log\int f_{\theta}(x,y)\log \left(\frac{f_{\theta}(x,y)}{\hat{I}(x,y)}\right)dxdy,\]

where $\hat{I}$ is defined as the image normalized to unit flux.

This struct is also a functor.

Extractor

This defines the interface to the optimizers that can find the optimal template for a given image.

VIDA.Optimizer — Type

abstract type Optimizer

Abstract optimizer class that defines the optimizer to use. This is the interface I'll need to define all the different optimizer classes below.

VIDA.BBO — Type

struct BBO <: VIDA.Optimizer

Defines the interface for the BlackBoxOptim interface. This requires that the user to have imported the BlackBoxOptim package.

Fields

popsize: The population size DEFAULT 64

maxevals: The maximum number of times to evaluate the divergence before terminiation.

tracemode: The output option. Default is silent, i.e. no output.

Other options are :compact and :verbose

Notes

This uses the default BlackBoxOptim optimizer i.e. adaptivederand1bin_radiuslimited. Currently other options aren't implemented since I found that this version tended to work the best.

VIDA.CMAES — Type

struct CMAES <: VIDA.Optimizer

Defines the interface for the LBFGS-B interface from the Optim.jl package. This requires that the user to have imported the CMAEvolutionStrategy package.

Typically I have found that this works very well. This the usual first optimizer to try.

Fields

popsize

PopulationSize Default: 64

cov_scale

Initial covariance scale Default: 1.0

ftarget

target divergence values. nothing means there is no target Default: nothing

xtol

xtol Default: nothing

ftol

ftol Default: 1.0e-11

maxevals

maximum number of divergence evals, nothing means run until termination Default: nothing

verbosity

verbosity of output Default: 1

VIDA.Opt — Type

struct Opt{O<:Optim.AbstractConstrainedOptimizer} <: VIDA.Optimizer

Defines the interface for the LBFGS-B interface from the Optim.jl package. This requires that the user to have imported the Optim package.

Fields

opt

options

Default: Optim.Options()

VIDA.ExtractProblem — Type

ExtractProblem{T<:AbstractDivergence, S<:AbstractTemplate}

Defines a feature extraction problem to minimize, with an abstract filte and an abstract divergence. This is needed to interface with the extractor minimizer, which will minimize the divergence to find the optimal template.

Fields

div

Divergence function to minimize

θinit

Initial location of the optimizer

θlower

Lower bound of the search region

θupper

Upper bound of the search region

VIDA.extractor — Function

extractor(prob::ExtractProblem, optimizer::Optimizer)

This extracts the optimal template defined by the prob problem. This will minimize the divergence in prob and return the optimal template and minimum divergence in a tuple.

optimizer is one of VIDA's optimizer types. Typically I would recommend the BBO() optimizer

Examples

    θopt, divmin = extractor(prob, BBO())

VIDA.threaded_extractor — Function

threaded_extractor(nstart::Int, prob::ExtractProblem, optimizer::Optimizer)

A threaded multi-start version of the extractor method. This will run nstart instances of extractor, where the initial location of chosen uniformly within the bounds defined in prob.

Outputs

This outputs the best template and minimum divergence of all the extractors run.

Utilities

VIDA.make_image — Function

make_image(θ, npix, xlim, ylim; intensity, source, wavelength, ra, dec, mjd)

Creates an EHTImage type from the template type. The number of pixels in the image are given by npix and the field of view in μas in the x and y direction are given by xlim and ylim. The rest of the options are the default image characteristics

make_image(θ, npix, xlim, ylim, source_img; intensity)

Creates an EHTImage type from the template type. The number of pixels in the image are given by npix and the field of view in μas in the x and y direction are given by xlim and ylim. To define the source we use an source_img.

Visualizations

VIDA.triptic — Function

triptic(img::EHTImage, template)

Plots a triptic where the left panel is the img middle the template and the right a two cross-sections of the image and template