API

VIDA.VIDA

VIDA.AbstractDivergence

VIDA.AbstractFilter

VIDA.AbstractFitsImage

VIDA.AbstractImage

VIDA.AddFilter

VIDA.AsymGaussian

VIDA.BBO

VIDA.Bhattacharyya

VIDA.CMAES

VIDA.Constant

VIDA.CosineRing

VIDA.CosineRing

VIDA.Disk

VIDA.EHTImage

VIDA.EllipticalGaussianRing

VIDA.ExtractProblem

VIDA.GaussianRing

VIDA.GeneralGaussianRing

VIDA.KullbackLeibler

VIDA.MulFilter

VIDA.Opt

VIDA.Optimizer

VIDA.SlashedGaussianRing

VIDA.TIDAGaussianRing

Base.split

VIDA.centroid

VIDA.clipimage

VIDA.downsample

VIDA.extractor

VIDA.field_of_view

VIDA.filter_image

VIDA.flux

VIDA.get_radec

VIDA.inertia

VIDA.load_ehtimfits

VIDA.load_fits

VIDA.rescale_image

VIDA.stack

VIDA.threaded_extractor

VIDA.unpack

VIDA.window_image

VIDA.VIDA — Module

VIDA

Is a image feature extraction tool for use with EHT images of black holes. It assumes that the image is close to one of the filters we have implemented and then tries to extract that feature from the image using one of the probability divergences implemented.

VIDA.AbstractDivergence — Type

abstract type AbstractDivergence

An abstract class for a divergence of a function. This is defined in terms of an Image type object. For two examples see below

VIDA.AbstractFilter — Type

abstract type AbstractFilter

An absract type that will contain the filter information, such as the parameters. Specific instanstantiations will need to be defined for you to use this.

Details

This defined the highest function type. If you wish to implement your own filter you
need to define a a couple of things
1. The filter type <: AbstractFilter
2. an functor of the type that computes the filter function
3. an `size` function that defines the number of parameters of the filter.

An example is given by:

#All of our composite type are defined using the Paramters.jl package to you
can directly refer to the struct parameters when creating it, although this isn't
actually used anywhere in the code.
@with_kw struct Gaussian <: AbstractFilter
    σ::Float64
    x0::Float64
    y0::Float64
end

#Typically we inline and force the function to use fastmath
@fastmath @inline function (θ::Gaussian)(x,y)
    return 1.0/(2π*σ^2)*exp(-0.5*( (x-x0)^2+(y-y0)^2 ) )
end
size(::Type{Gaussian}) = 3

VIDA.AbstractFitsImage — Type

An absract image that will hold a fits image after being created or parsed in. This will form the basis for most astronomical images that are defined.

VIDA.AbstractImage — Type

An abstact type that acts as a wrapper for image objects used in astronomy.

This is the top of the castle for images and will be rarely touched. Basically unless you don't want to use fits images this will not be used

VIDA.AddFilter — Type

struct AddFilter{T1<:VIDA.AbstractFilter, T2<:VIDA.AbstractFilter} <: VIDA.AbstractFilter

Combines two filters together into one object. Since addition is assoiciative this can actually we used to hold multiple different filters.

Details

Overloads the Base.:+ function so you can easily add two filters together.

Example

θ1 = GaussianRing(10,5,0,0)
θ2 = SlashedGaussianRing(15,5,0.5,π/4,0,0)
θ12 = θ1+θ2

VIDA.AsymGaussian — Type

struct AsymGaussian <: VIDA.AbstractFilter

An asymmetric Gaussian blob.

Details

Defines a asymmetric Gaussian image. This is useful if the image has some non-ring emission in the and you need to soak up some of the flux.

The parameters of the model follow very closely to those used in Themis. The Gaussian size σ is given by σ = √(σxσy) τ = 1-σy/σx, where σx,σy are the semi-major,minor axis lenght respectively. This is similar to how the asymmetry for the EllipticalGaussianRing.

Fields

σ

Gaussian size in μas

τ

Gaussian asymmetry

ξ

Gaussian orientation in radians measured north of east

x0

x position of Gaussian center in μas

y0

y position of Gaussian center in μas

VIDA.BBO — Type

struct BBO <: VIDA.Optimizer

Defines the interface for the BlackBoxOptim interface. This requires that the user to have imported the BlackBoxOptim package.

Fields

popsize: The population size DEFAULT 64

maxevals: The maximum number of times to evaluate the divergence before terminiation.

tracemode: The output option. Default is silent, i.e. no output.

Other options are :compact and :verbose

Notes

This uses the default BlackBoxOptim optimizer i.e. adaptivederand1bin_radiuslimited. Currently other options aren't implemented since I found that this version tended to work the best.

Fields

method

Default: :adaptivederand1bin_radiuslimited

popsize

Default: 64

maxevals

Default: 25000

tracemode

Default: :compact

VIDA.Bhattacharyya — Type

struct Bhattacharyya{T, S} <: VIDA.AbstractDivergence

Type for the Bhattacharyya divergence. It constructed from an EHTImage i.e. data. Additionally to evaluate the divergence we use a functor approach where if θ is your

Details

This computes the Bhattacharyya divergence which is related to Hellinger distance between two distributions. In fact, they are both minimized at the same point. The Bhattacharyya divergence is defined as

\[Bh(f_\theta||\hat{I}) = -\log\int \sqrt{f_\theta(x,y)\hat{I}(x,y)}dxdy,\]

where $\hat{I}$ is defined as the image normalized to unit flux.

VIDA.CMAES — Type

struct CMAES <: VIDA.Optimizer

Defines the interface for the LBFGS-B interface from the Optim.jl package. This requires that the user to have imported the CMAEvolutionStrategy package.

Typically I have found that this works very well. This the usual first optimizer to try.

Fields

popsize

PopulationSize Default: 64

cov_scale

Initial covariance scale Default: 1.0

ftarget

target divergence values. nothing means there is no target Default: nothing

xtol

xtol Default: nothing

ftol

ftol Default: 1.0e-11

maxevals

maximum number of divergence evals, nothing means run until termination Default: nothing

verbosity

verbosity of output Default: 1

VIDA.Constant — Type

struct Constant <: VIDA.AbstractFilter

An constant filter.

Details

Defines an image that just has constant flux. This is very useful for soaking up low levels of flux in image reconstructions that can bias the results.

Since images or normalized to unity, this means the Constant filter has no additional parameters.

VIDA.CosineRing — Type

struct CosineRing{N, M} <: VIDA.AbstractFilter

Extrememly flexible ring model. The thickness is modeled as a cosine expansion with N terms and the slash by a expansion with M terms.

Details

The ring is allowed to be elliptical. The thickness of the ring is modeled by a cosine expansion in azimuthal angle. N specifies the number of cosine modes to fit, where the first mode is the constant thickness portion and so has no corresponding angle. The slash is modeled as a separate cosine expansion, with M terms. Here the zero order term is forced to be unity, so M defines the M additional terms.

VIDA.CosineRing — Method

CosineRing{N,M}(p::AbstractArray) where {N,M}

Takes in a vector of paramters describing the filter.

Details

The order of the vector must be

p[1] = r0

p[2:(N+1)] = σ

p[(N+2):(2N)] = ξσ

p[2N+1] = τ

p[2N+2] = ξτ

p[2N+3:2N+M+2] = s

p[2N+3+M:2N+2+2M] = ξs

p[2N+3+2M] = x0

p[2N+4+2M] = y0

VIDA.Disk — Type

struct Disk <: VIDA.AbstractFilter

A smoothed disk model

Details

Defines a filter for an image that has a smoothed disk model.

VIDA.EHTImage — Type

struct EHTImage{T} <: VIDA.AbstractFitsImage{T}

Details

The trait is to hold a EHT image tyically in matrix form. Namely the trait will typically be Matrix{Float64}.

nx is the number of pixels in the x or RA direction ny is the number of pixels in the y or DEC direction psize_x, psize_y are the pixel sizes in the x and y direction source is the source we are looking at e.g. M87 ra,dec are the sources RA and DEC in J2000 coordinates using degrees wavelength is the wavelength of the image. mjd is the Modified Julian Date of the observation. img is the actual pixeled image in Jy/pixel

VIDA.EllipticalGaussianRing — Type

struct EllipticalGaussianRing <: VIDA.AbstractFilter

Implements the elliptical gaussian ring filter. Where the ellipticity tau is defined as one minus ratio between the semi-minor and semi-major axis.

Details

Adds ellipticity to the ring. The radius r0 of the ring is now defined as the geometric mean of the semi-major (a) and semi-minor (b) axis lengths i.e.

r0 = √(a*b).

The ellipticity τ is given by τ = 1-b/a.

Fields

r0

Radius of the Gaussian ring

σ

Standard deviation of the width of the Gaussian ring

τ

Asymmetry of the Gaussian ring defined as $1-b/a$

ξ

Asymmetry orientation in radians measured north of east

x0

x position of the center of the ring in μas

y0

y position of the center of the ring in μas

Notes

There is no normalization since the ellipticity makes it impossible to normalize analytically. In fact the distance from the ellipse is implemented numerically using an algorithm adapted from git

VIDA.ExtractProblem — Type

ExtractProblem{T<:AbstractDivergence, S<:AbstractFilter}

Defines a feature extraction problem to minimize, with an abstract filte and an abstract divergence. This is needed to interface with the extractor minimizer, which will minimize the divergence to find the optimal filter.

Fields

div

Divergence function to minimize

θinit

Initial location of the optimizer

θlower

Lower bound of the search region

θupper

Upper bound of the search region

VIDA.GaussianRing — Type

struct GaussianRing <: VIDA.AbstractFilter

Symmetric gaussian ring filter. This is the most basic filter and just attempts to recover a location x0,y0, radius r0 and thickness σ from some image.

Fields

r0

Radius of Gaussian ring in μas

σ

Standard deviation of Gaussian ring in μas

x0

x location of the center of the Gaussian ring in μas

y0

y location of the center of the Gaussian ring in μas

Example

GaussianRing(r0=20.0,σ=5.0,x0=0.0,y0=-10.0)

VIDA.GeneralGaussianRing — Type

struct GeneralGaussianRing <: VIDA.AbstractFilter

Creates the most general elliptical slashed gaussian ring model. It is a combination of the elliptical and slashed gaussian ring. The direction of the slash and the ellipticity are not aligned or anti-aligned like with the TIDAGaussianRing type.

Details

Adds ellipticity to the ring. The radius r0 of the ring is now defined as the geometric mean of the semi-major (a) and semi-minor (b) axis lengths i.e.

r0 = √(a*b).

The ellipticity τ is given by τ = 1-b/a.

Fields

r0

Radius of the Gaussian ring

σ

Standard deviation of the width of the Gaussian ring

τ

Asymmetry of the Gaussian ring defined as $1-b/a$

ξτ

Asymmetry orientation in radians, measured north of east

s

Slash of Gaussian ring.

ξs

Slash orientation in radians measured north of east

x0

x position of the center of the ring in μas

y0

y position of the center of the ring in μas

VIDA.KullbackLeibler — Type

struct KullbackLeibler{T, S} <: VIDA.AbstractDivergence

Type for the KL divergence. It constructed from an EHTImage i.e. data. Additionally to evaluate the divergence we use a functor approach where if θ is your

Details

This computes the KL divergence which is related to Hellinger distance between two distributions. In fact, they are both minimized at the same point. The Bhattacharyya divergence is defined as

\[KL(f_\theta||\hat{I}) = -\log\int f_{\theta}(x,y)\log \left(\frac{f_{\theta}(x,y)}{\hat{I}(x,y)}\rightdxdy,\]

where $\hat{I}$ is defined as the image normalized to unit flux.

This struct is also a functor. ```

VIDA.MulFilter — Type

struct MulFilter{T<:VIDA.AbstractFilter, S<:Number} <: VIDA.AbstractFilter

Multiplies filter by a constant. This is useful when combining with AddFilter since it will change the relative weights of each filter.

Details

Overloads the Base.:* function so you can easily multiple a filter by a number.

Example

θ = GaussianRing(15,5,0.0,0.0)
2*θ

VIDA.Opt — Type

struct Opt{O<:Optim.AbstractConstrainedOptimizer} <: VIDA.Optimizer

Defines the interface for the LBFGS-B interface from the Optim.jl package. This requires that the user to have imported the Optim package.

Fields

opt

options

Default: Optim.Options()

VIDA.Optimizer — Type

abstract type Optimizer

Abstract optimizer class that defines the optimizer to use. This is the interface I'll need to define all the different optimizer classes below.

VIDA.SlashedGaussianRing — Type

struct SlashedGaussianRing <: VIDA.AbstractFilter

Implements the slashed gaussian ring filter, that uses a cosine to symmetrically implement the slash. While this is marginally more complicated that a linear slash, it has a number of benefits such as mainting the azimuthal and smooth structure of the image.

Fields

r0

Radius of the ring in μas

σ

Standard deviation of Gaussian ring in μas

s

Slash strength of Gaussiang ring. 0 means no slash

ξ

Rotation angle in radians of slash direction, measured north of west

x0

x position of the center of the ring in μas

y0

y position of the center of the ring in μas

VIDA.TIDAGaussianRing — Type

struct TIDAGaussianRing <: VIDA.AbstractFilter

Creates the filter from the Paper I am writing. It is a combination of the elliptical and slashed gaussian ring. The slash and the semi-major axis are either aligned if the slash parameter s>0 or antialigned if s<0.

Details

Adds ellipticity to the ring. The radius r0 of the ring is now defined as the geometric mean of the semi-major (a) and semi-minor (b) axis lengths i.e.

r0 = √(a*b).

The ellipticity τ is given by τ = 1-b/a.

Fields

r0

Radius of the Gaussian ring

σ

Standard deviation of the width of the Gaussian ring

τ

Asymmetry of the Gaussian ring defined as $1-b/a$

s

Slash of Gaussian ring.

ξ

Slash/Asymmetry orientation in radians measured north of east

x0

x position of the center of the ring in μas

y0

y position of the center of the ring in μas

Base.split — Method

split(θ)

Splits the filter into an array with its subcomponents so you can easily access them.

VIDA.centroid — Method

centroid(img)

Finds the centroid or center of light of the img in μas.

VIDA.clipimage — Function

clipimage(clip, im)
clipimage(clip, im, mode)

Clips the image im according to the value clip, which can either be an absolute flux in Jy/px or the intensity relative to the maximum.

VIDA.downsample — Method

downsample(factor, im)

Down samples the image im by a factor

Details

Given an image im with $Nx\times Ny$ pixels, downsample converts that to an image with $Nx/factor \times Ny/factor$ pixels. This can be useful when the image has much higher resolutions than is needed for feature extraction.

VIDA.extractor — Method

extractor(prob::ExtractProblem, optimizer::Optimizer)

This extracts the optimal filter defined by the prob problem. This will minimize the divergence in prob and return the optimal filter and minimum divergence in a tuple.

optimizer is one of VIDA's optimizer types. Typically I would recommend the BBO() optimizer

Examples

    θopt, divmin = extractor(prob, BBO())

VIDA.field_of_view — Method

field_of_view(img::EHTImage)

Finds the field of view of an EHTImage. Return a w element tuple with the field of view in the x and y direction

VIDA.filter_image — Method

filter_image(θ, npix, xlim, ylim)

Creates an npix×npix rasterized image of the filter θ with limits xlim and ylim

Returns the tuple (xitr,yitr,image) where xitr,yitr are the iterators defining the pixel locations (which are centered) and the rasterized image, in Jy/μas^2.

Note

I use the pixel size definition fieldofview/npix, but the image is evaluated at the pixel centers.

We also use the astronomer orientation and ordering.

VIDA.flux — Method

flux(img)

Finds the image flux of an EHTImage img

VIDA.get_radec — Method

get_radec(img::T) <: AbstractFitsImage

Returns two iterators (ra,dec) that give the locations of the img pixels.

VIDA.inertia — Function

inertia(img)
inertia(img, center)

Find the image moment of inertia or second moment

Notes

If center=true then we find the central second moment, or the second cumulant of the image.

VIDA.load_ehtimfits — Method

load_ehtimfits(fits_name)

where fits_name should be a fits file generated using ehtim

Details

This reads in a fits file created using ehtim. This is because ehtim only outputs the image and not a separate HDU for the field, so the usual fits reader doesn't work properly.

The function returns an EHTImage object that contains the relevant image and parameters extracted from the fits file. It also ensures that we are astronomers and that the image using sky-left coordinates.

VIDA.load_fits — Method

load_fits(fits_name)

where fits_name should be a fits file generated using ehtim

Details

This reads in a fits file that is more robust to the various imaging algorithms in the EHT, i.e. is works with clean, smili, eht-imaging.

The function returns an EHTImage object that contains the relevant image and parameters extracted from the fits file. It also ensures that we are astronomers and that the image using sky-left coordinates.

VIDA.rescale_image — Method

rescale_image(img::EHTImage, npix, xlim, ylim)

Inputs

img::EHTImage : Image you want to rescale

npix : Number of pixels in x and y direction

xlim : Tuple with the limits of the image in the RA

ylim : Tuple with the limits of the image in DEC

VIDA.stack — Method

stack(θ, θ1)

Stacks filters together so you can easily combine multiple filters. It does this by calling the :+ and :* method. Every filter added will include an additional parameter that controls the relative weight of each filter.

VIDA.threaded_extractor — Method

threaded_extractor(nstart::Int, prob::ExtractProblem, optimizer::Optimizer)

A threaded multi-start version of the extractor method. This will run nstart instances of extractor, where the initial location of chosen uniformly within the bounds defined in prob.

Outputs

This outputs the best filter and minimum divergence of all the extractors run.

VIDA.unpack — Method

unpack(θinit)

Unpacks the parameters of the filter θ

Returns the parameters in a vector.

VIDA.window_image — Method

window_image(domain_x, domain_y, im)

Given an image im it selects a new field of view for the image given by domain_x, domain_y.