API

Contents

API

Contents

Index

Index

HypercubeTransform.ascube

HypercubeTransform.asflat

HypercubeTransform.transform_tuple

TransformVariables.inverse

TransformVariables.transform

TransformVariables.transform

HypercubeTransform.ascube — Function

`ascube(c)

Constructs the object that contains the necessary information to move from the unit hypercube to the distribution space. This is the usual function to use when construct the transformation.

There are a few different behaviors depending on the type of the object.

If c::Distribution then this will store the distributions.

If c::Tuple{AbstractHypercubeTransform} then this will store the tuple

Examples

ascube(Normal())
ascube(MultivariateNormal())
ascube((Normal(), Normal(2.0)))
ascube( (α = Uniform(), β = Normal()) )

HypercubeTransform.asflat — Function

asflat(d::Distribution)

Computes the transformation of the support of the distribution d such that the variables live on ℝⁿ where n is the dimension of the problem. This is essentially what Turing and Stan do when reparameterizing the model.

The returned object is a TransformVariables.AbstractTransform object and follows that interface. Please see the TransformVariable docs for more information.

HypercubeTransform.transform_tuple — Method

transform_tuple(tt, x, index)

Helper function that steps through the transformation tuple

TransformVariables.inverse — Method

`inverse(c::AbstractHypercubeTransform, p)`

Transforms from the parameter space p, to the unit hypercube defined by the transformation c.

The behavior of this function depends on the nature of c.

If c is a <: Distributions.Distributions and has a cdf method

this will just call the cdf function. If no cdf function is defined then a custom transformation depending on the type of c will be called. If no custom transformation exists then an error will be raised.

If c is a Tuple of transformations then inverse will iterate through the

tuple using a similar method to the TransformVariables.jl method.

TransformVariables.transform — Function

transform(c::AbstractHypercubeTransform, p)

Transforms from the hypercube with coordinates p, to the parameter space defined by the transformation c.

The behavior of this function depends on the nature of c.

If c is a <: Distributions.Distributions and has a quantile method

this will just call the quantile function. If no quantile function is defined then a custom transformation depending on the type of c will be called. If no custom transformation exists then an error will be raised.

If c is a Tuple of transformations then transform will iterate through the

tuple using a similar method to the TransformVariables.jl method.

TransformVariables.transform — Method

transform(c, x)

Computes the transformation from the unit hypercube to the distribution space.