API

Contents

<!– ```@contents Pages = ["function_index.md"]

## Index

<!-- ```@index
Pages = ["function_index.md"]

–>

VIDA.VIDA — Module

VIDA

Is a image feature extraction tool for use with EHT images of black holes. It assumes that the image is close to one of the templates we have implemented and then tries to extract that feature from the image using one of the probability divergences implemented.

VIDA.AbstractDivergence — Type

abstract type AbstractDivergence

Defines an AbstractDivergence type. This is the basic cost function of VIDA.

A subtype of AbstractDivergence must be a struct with at least two fields

img::IntensityMap which holds the image you wish to run feature extraction on

mimg::IntensityMap which is an internal buffer that stores the model image.

The divergence is evaluated roughly as

    normalize_div(div, sum(divergence_point(div, image, model)))

Therefore a user must implement the following methods

divergence_point: Which evaluates the divergence at a single pixel

normalize_div: Which normalizes and modifies the

<!– @autodocs Modules = [VIDA] –>