API

Contents

API

Contents

Index

Index

VIDA.VIDA

VIDA.AbstractDivergence

VIDA.AbstractImageTemplate

VIDA.AbstractMovie

VIDA.AzimuthalCosine

VIDA.AzimuthalUniform

VIDA.Bhattacharyya

VIDA.Constant

VIDA.CosineRing

VIDA.GaussDisk

VIDA.GaussDisk

VIDA.KullbackLeibler

VIDA.LeastSquares

VIDA.LogSpiral

VIDA.LogSpiral

VIDA.RadialDblPower

VIDA.RadialGaussian

VIDA.RadialTruncExp

VIDA.Renyi

VIDA.RingTemplate

VIDA.VIDAMovie

VIDA.VIDAMovie

VIDA.VIDAProblem

VIDA.VIDAProblem

ComradeBase.flux

VIDA.CosineRingwFloor

VIDA.CosineRingwGFloor

VIDA.EllipticalCosineRing

VIDA.EllipticalGaussianRing

VIDA.EllipticalSlashedGaussianRing

VIDA.GaussianRing

VIDA.GaussianRing

VIDA.SlashedGaussianRing

VIDA.SlashedGaussianRing

VIDA.blur

VIDA.blur

VIDA.clipimage

VIDA.divergence

VIDA.divergence_point

VIDA.get_frames

VIDA.get_image

VIDA.get_times

VIDA.join_frames

VIDA.load_hdf5

VIDA.load_im_h5

VIDA.normalize_div

VIDA.save_hdf5

VIDA.triptic

VIDA.vida

VLBISkyModels.regrid

VIDA.VIDA — Module

VIDA

Is a image feature extraction tool for use with EHT images of black holes. It assumes that the image is close to one of the templates we have implemented and then tries to extract that feature from the image using one of the probability divergences implemented.

VIDA.AbstractDivergence — Type

abstract type AbstractDivergence

Defines an AbstractDivergence type. This is the basic cost function of VIDA.

A subtype of AbstractDivergence must be a struct with at least two fields

img::IntensityMap which holds the image you wish to run feature extraction on

mimg::IntensityMap which is an internal buffer that stores the model image.

The divergence is evaluated roughly as

    normalize_div(div, sum(divergence_point(div, image, model)))

Therefore a user must implement the following methods

divergence_point: Which evaluates the divergence at a single pixel

normalize_div: Which normalizes and modifies the

```

VIDA.AbstractImageTemplate — Type

abstract type AbstractImageTemplate <: ComradeBase.AbstractModel

An abstract ComradeBase.AbstractModel that serves as the parent type of all the VIDA templates implemented in this repo.

By default this model assumes the following ComradeBase traits

ComradeBase.visanalytic(::Type{<:AbstractImageTemplate}) = NoAnalytic()
ComradeBase.imanalytic(::Type{<:AbstractImageTemplate})  = IsAnalytic()
ComradeBase.ispolarized(::Type{<:AbstractImageTemplate}) = NotPolarized()

As a result if a user wishes the implement their own subtype (e.g., MyTemplate) of AbstratImageTemplate they will need to implement the following methods

ComradeBase.intensity_point(m::MyTemplate, p): which computes the potentially unormalized brightness of the template at the point p.

ComradeBase.radialextent(m::MyTemplate): which computes the rough radial extent of the model m.

For more information about the total interface see VLBISkyModels.jl

VIDA.AbstractMovie — Type

abstract type AbstractMovie

Creates an abstract movie class

VIDA.AzimuthalCosine — Type

AzimuthalCosine(s::NTuple{N,T}, ξ::NTuple{N, T}) where {N, T}

A azimuthal profile that is given by a cosine expansion of order N. The expansion is given by

    1 - ∑ₙ sₙcos(nϕ - ξₙ)

Notes

This is usually couple with a radial profile to create a general ring template

julia> rad = RadialDblPower(3.0, 3.0)
julia> azi = AzimuthalCosine((0.5, 0.2), (0.0, π/4)) # Defaults to Float64
julia> t = RingTemplate(rad, azi)

Arguments

s : amplitudes of the N order cosine expansion of the azimuthal brightness

ξs: phase of the N order cosine expansion of the azimuthal brightness

VIDA.AzimuthalUniform — Method

AzimuthalUniform()
AzimuthalUniform{T}()

A azimuthal profile that is uniform for all angles.

Notes

This is usually couple with a radial profile to create a general ring template

julia> rad = RadialDblPower(3.0, 3.0)
julia> azi = AzimuthalUniform() # Defaults to Float64
julia> t = RingTemplate(rad, azi)

VIDA.Bhattacharyya — Type

struct Bhattacharyya{T<:(KeyedArray{T, N, <:Any, G} where {T, N, G<:ComradeBase.AbstractDims})} <: VIDA.AbstractDivergence

Type for the Bhattacharyya divergence. It constructed from an EHTImage i.e. data. Additionally to evaluate the divergence we use a functor approach where if θ is your

Details

This computes the Bhattacharyya divergence which is related to Hellinger distance between two distributions. In fact, they are both minimized at the same point. The Bhattacharyya divergence is defined as

\[Bh(f_\theta||\hat{I}) = -\log\int \sqrt{f_\theta(x,y)\hat{I}(x,y)}dxdy,\]

where $\hat{I}$ is defined as the image normalized to unit flux.

VIDA.Constant — Type

struct Constant{T} <: VIDA.AbstractImageTemplate

An constant template.

Details

Defines an image that just has constant flux. This is very useful for soaking up low levels of flux in image reconstructions that can bias the results.

Since images or normalized to unity, this means the Constant template has no additional parameters.

Fields

scale: Sets the angular scale of the image. This is usually equal to the image FOV

VIDA.CosineRing — Type

struct CosineRing{N, M, T} <: VIDA.AbstractImageTemplate

Extrememly flexible symmetric ring model. The thickness is modeled as a cosine expansion with N terms and the slash by a expansion with M terms.

Details

The ring is forced to be symmetric for a significant speed boost over CosineRing. The thickness of the ring is modeled by a cosine expansion in azimuthal angle. N specifies the number of cosine modes to fit, where the first mode is the constant thickness portion and so has no corresponding angle. The slash is modeled as a separate cosine expansion, with M terms. Here the zero order term is forced to be unity, so M defines the M additional terms.

VIDA.GaussDisk — Type

struct GaussDisk{T} <: VIDA.AbstractImageTemplate

A smoothed disk model

Details

Defines a template for an image that has a smoothed disk model.

Fields

α: Disk edge standard deviation

VIDA.GaussDisk — Method

GaussDisk(r0, σ, x0, y0)

Constructs a disk template with a Gaussian fall off.

Notes

This is a convienence constructor that uses VLBISkyModels modify under the hood. To create this function yourself do

modify(GaussDisk(σ/r0), Stretch(r0), Shift(x0, y0))

Arguments

r0: Radius of flat top of the disk.

σ: Standard deviation of the Gaussian of the disk edge.

x0: The horizontal location of the disk center

y0: The vertical location of the disk center

VIDA.KullbackLeibler — Type

struct KullbackLeibler{T<:(KeyedArray{T, N, <:Any, G} where {T, N, G<:ComradeBase.AbstractDims})} <: VIDA.AbstractDivergence

Type for the KL divergence. It constructed from an EHTImage i.e. data. Additionally to evaluate the divergence we use a functor approach where if θ is your

Details

This computes the KL divergence which is related to Hellinger distance between two distributions. In fact, they are both minimized at the same point. The Bhattacharyya divergence is defined as

\[KL(f_\theta||\hat{I}) = -\log\int f_{\theta}(x,y)\log \left(\frac{f_{\theta}(x,y)}{\hat{I}(x,y)}\right)dxdy,\]

where $\hat{I}$ is defined as the image normalized to unit flux.

This struct is also a functor.

VIDA.LeastSquares — Type

struct LeastSquares{T} <: VIDA.AbstractDivergence

Type for the least squares divergence. It constructed from an EHTImage i.e. data. Additionally to evaluate the divergence we use a functor approach where if θ is your

Details

This computes the squared 2 norm between your image and template, both of which are normalized to unit flux.

To construct this just pass it an image object

julia> ls = LeastSquares(img::EHTImage)

Notes

We have a template internal matrix the prevents internal allocations during computation This is a internal feature that the user does not need to worry about.

VIDA.LogSpiral — Type

struct LogSpiral{T<:Real} <: VIDA.AbstractImageTemplate

Template type for a logarithmic spiral segment

Fields

κ: Unit curvature of the logarithmic spiral

σ: thickness of the Gaussian spiral arm

δϕ: Azimuthal extent of the spiral arm

VIDA.LogSpiral — Method

LogSpiral(r0, κ, σ, δϕ, ξ, x0, y0)

Constructs a LogSpiral template.

Notes

This is a convienence constructor that uses VLBISkyModels modify under the hood. To create this function yourself do

modify(LogSpiral(κ, σ/r0, δϕ),
       Stretch(r0), Rotate(ξ), Shift(x0, y0))

Arguments

r0: Radius of the start of the logarithmic spiral

κ: Unit curvature of the spiral

σ: Thickness of the logarithmic spiral

δϕ: The angular extent of the spiral

ξ: The position angle of the start of the spiral

x0: The horizontal location of the start of the spiral

y0: The vertical location of the start of the spiral

VIDA.RadialDblPower — Type

RadialDblPower(αinner, αouter)

Radial profile that given by a double power-law with the function form

    r^αinner*inv(1+ r^(αinner + αouter + 1))

Notes

This is usually couple with a azimuthal profile to create a general ring template

julia> rad = RadialDblPower(3.0, 3.0)
julia> azi = AzimuthalUniform()
julia> t = RingTemplate(rad, azi)

Arguments

αinner the power law index for r<1.

αouter the power law index for r≥1.

VIDA.RadialGaussian — Type

RadialGaussian(σ)

Create a profile that is a radial gaussian ring with radius unity and standard deviation σ.

Notes

This is usually couple with a azimuthal profile to create a general ring template

julia> rad = RadialGaussian(0.1)
julia> azi = AzimuthalUniform()
julia> t = RingTemplate(rad, azi)

Arguments

σ: The standard deviation for the Gaussian ring.

VIDA.RadialTruncExp — Type

RadialTruncExp(σ)

Radial profile that has a exponential profile up to a cutoff radius of 1 where for r<1 the profile is identically zero.

This evalutes to

    exp(-(r-1)/σ)

Notes

This is usually couple with a azimuthal profile to create a general ring template

julia> rad = RadialTruncExp(2.0)
julia> azi = AzimuthalUniform()
julia> t = RingTemplate(rad, azi)

Arguments

σ: Exponential inverse fall off parameter.

VIDA.Renyi — Method

Renyi(img::EHTImage, α)

Construct the Renyi divergence with parameter α. It constructed from an EHTImage i.e. data.

Details

This computes the KL divergence which is related to Hellinger distance between two distributions. In fact, they are both minimized at the same point. The Bhattacharyya divergence is defined as

\[Ry(f_\theta||\hat{I}) = \frac{1}{α-1}\log\int log \left(\frac{f_{\theta}(x,y)^\alpha}{\hat{I}(x,y)^{\alpha-1}}\right)dxdy,\]

where $\hat{I}$ is defined as the image normalized to unit flux.

This is a very flexible divergence that reduces to many of the other divergences implemented.

α = 1 corresponds to the KL divergence

α = 1/2 corresponds to the Bhattacharyya divergence up to a multiplicative factor of 2

Typically we find that α=1.5 works well, as it focusses on the bright regions of the images moreso than the Bh and KL divergence. For α>2 the measure tends to devolve in something akin the to sup norm and fails to match the image structure.

VIDA.RingTemplate — Type

struct RingTemplate{R<:VIDA.AbstractRadial, A<:VIDA.AbstractAzimuthal} <: VIDA.AbstractImageTemplate

A flexible ring template that forms a ring by taking the product of a radial and azimuthal brightness profile.

A list of radial profiles is given by subtypes(AbstractRadial)

A list of azimuthal profiles is given by subtypes(AbstractAzimuthal)

Examples

julia> rad = RadialGaussian(0.1)
julia> azi = AzimuthalUniform()
julia> ring = modify(RingTemplate(rad, azi), Stretch(10.0), Shift(1.0, 2.0))

Fields

radial: Radial profile of the ring

azimuthal: Azimuthal profile of the ring

VIDA.VIDAMovie — Type

struct VIDAMovie{T, F<:(KeyedArray{T, 3, <:Any, G} where G<:ComradeBase.AbstractDims), I<:Interpolations.AbstractExtrapolation} <: VIDA.AbstractMovie

Details

Holds a X,Y,T IntensityMap plus an interpolator that lets you make a continuous movie

VIDA.VIDAMovie — Method

VIDAMovie(mov::IntensityMap{T,3})

Create an VIDAMovie class for easy interpolation between frames.

Arguments

mov: An IntensityMap with axes (:X, :Y, :T) that represent the frames of a movie. Note that the time dimension does not have to be equi-spaced.

Returns

An VIDAMovie object that behaves like IntensityMap but lets you interpolate between frames with get_image(vida_movie, time).

VIDA.VIDAProblem — Type

struct VIDAProblem{D<:VIDA.AbstractDivergence, F, N, B}

A composite type that holds various properties that define the optimization process to extract the optimal filter.

Fields

div: Divergence you will fit

f: Function that defines the parametric family of templates

autodiff: Type of autodiff to use when optimizing if any

lb: The lower bounds of the parameter ranges to search over

ub: The upper bounds of parameter ranges to search over

VIDA.VIDAProblem — Method

VIDAProblem(div, f, lb, ub)

Defines a VIDAProblem for optimization.

Arguments

div: The divergence you wish to fit. This is an instantiation of VIDA.AbstractDivergence

f: The function that defines the parametric family of templates you will fit. The function must accept a named tuple as its first argument, whose names define the parameters.

lb: A NamedTuple whose names match the argument of f and whose values define the lower bounds of the parameters range you want to search over

lb: A NamedTuple whose names match the argument of f and whose values define the upper bounds of the parameters range you want to search over

Example

julia> f(x) = SlashedGaussianRing(x.σ, x.s)
julia> div = Renyi(img, 0.75)
julia> lb = (x = 0.1, s = 0.001)
julia> ub = (x = 10.0, s = 0.999)
julia> prob = VIDAProblem(div, f, lb, ub)

ComradeBase.flux — Method

flux(mov, t)

Returns the flux of the mov at the times time in fractional hours

VIDA.CosineRingwFloor — Method

CosineRingwFloor(r0, σ0, σ, ξσ, s, ξs, floor, x0, y0)

A Cosine ring with a GaussDisk at the center of the ring that matches the r0 and σ0 of the CosineRing.

This is a convienence constructor of the more basic VLBISkyModel image constructors. It is equivalent to

julia> CosineRing(r0, σ0, σ, ξσ, s, ξs, x0, y0) + floor*modify(GaussDisk(σ0/r0), Stretch(r0), Shift(x0, y0))

Details

Adds ellipticity to the ring. The radius r0 of the ring is now defined as the geometric mean of the semi-major (a) and semi-minor (b) axis lengths i.e.

r0 = √(a*b).

The ellipticity τ is given by τ = 1-b/a.

Arguments

r0: geometric mean radius (√ab) of the ring

σ0: standard deviation of the Gaussian ring

σ : amplitudes of the N order cosine expansion of the ring thickness

ξσ: phase of the N order cosine expansion of the ring thickness

s : amplitudes of the M order cosine expansion of the azimuthal brightness

ξs: phase of the M order cosine expansion of the azimuthal brightness

floor: The flux of the center Gaussian. This is relative to the CosineRing.

x0: location of the ring center horizontally

y0: location of the ring center vertically

VIDA.CosineRingwGFloor — Method

CosineRingwGFloor(r0, σ0, σ, ξσ, s, ξs, floor, σG, x0, y0)

A Cosine ring with a Gaussian blob at the center of the ring. This is a convienence constructor of the more basic VLBISkyModel image constructors. It is equivalent to

julia> CosineRing(r0, σ0, σ, ξσ, s, ξs, x0, y0) + floor*modify(Gaussian(), Stretch(σG), Shift(x0, y0))

Details

Adds ellipticity to the ring. The radius r0 of the ring is now defined as the geometric mean of the semi-major (a) and semi-minor (b) axis lengths i.e.

r0 = √(a*b).

The ellipticity τ is given by τ = 1-b/a.

Arguments

r0: geometric mean radius (√ab) of the ring

σ0: standard deviation of the Gaussian ring

σ : amplitudes of the N order cosine expansion of the ring thickness

ξσ: phase of the N order cosine expansion of the ring thickness

s : amplitudes of the M order cosine expansion of the azimuthal brightness

ξs: phase of the M order cosine expansion of the azimuthal brightness

floor: The flux of the center Gaussian. This is relative to the CosineRing.

σG: The size of the central Gaussian.

x0: location of the ring center horizontally

y0: location of the ring center vertically

VIDA.EllipticalCosineRing — Method

EllipticalCosineRing(r0, σ0, σ, ξσ, τ, ξτ, s, ξs, x0, y0)

An Elliptical Cosine ring. This is a convienence constructor of the more basic VLBISkyModel image constructors. It is equivalent to

julia> modify(CosineRing(σ0/r0, σ/r0, ξσ, s, ξs .- ξτ),
            Stretch(r0*sqrt(1-τ), r0/sqrt(1-τ)),
            Rotate(ξτ),
            Shift(x0, y0)
        )

Details

Adds ellipticity to the ring. The radius r0 of the ring is now defined as the geometric mean of the semi-major (a) and semi-minor (b) axis lengths i.e.

r0 = √(a*b).

The ellipticity τ is given by τ = 1-b/a.

Arguments

r0: geometric mean radius (√ab) of the ring

σ0: standard deviation of the Gaussian ring

σ : amplitudes of the N order cosine expansion of the ring thickness

ξσ: phase of the N order cosine expansion of the ring thickness

τ : asymmetry of the ring τ = 1-b/a

ξτ: semi-major axis measured east of north

s : amplitudes of the M order cosine expansion of the azimuthal brightness

ξs: phase of the M order cosine expansion of the azimuthal brightness

x0: horizontal location of the ring center

y0: vertical location of the ring center vertically

VIDA.EllipticalGaussianRing — Method

EllipticalGaussianRing(r0, σ, τ, ξτ, x0, y0)

Implements the elliptical gaussian ring template with semi-minor axis a and semi-major axis b.

Arguments

r0: geometric mean radius (√ab) of the ring

σ : standard deviation of the Gaussian ring

τ : asymmetry of the ring τ = 1-b/a

ξτ: semi-major axis measured east of north

x0: location of the ring center horizontally

y0: location of the ring center vertically

Notes

This is a convienence constructor that uses RingTemplate under the hood. To create this function your self do

julia modify(GaussianRing(σ/r0), Stretch(r0*sqrt(1-τ), r0/sqrt(1-τ)), Rotate(ξτ), Shift(x0, y0))

VIDA.EllipticalSlashedGaussianRing — Method

EllipticalSlashedGaussianRing(r0, σ, τ, ξτ, s, ξs, x0, y0)

Combination of the elliptical and slashed gaussian ring.

Details

Adds ellipticity to the ring. The radius r0 of the ring is now defined as the geometric mean of the semi-major (a) and semi-minor (b) axis lengths i.e.

r0 = √(a*b).

The ellipticity τ is given by τ = 1-b/a.

The brightness asymmetry uses a cosine to implement the slash.

Arguments

r0: geometric mean radius (√ab) of the Gaussian ring

σ : standard deviation of the Gaussian ring

τ : asymmetry of the Gaussian ring τ = 1-b/a

ξτ: semi-major axis measured east of north

s : Slash amplitude. 0 means no slash, and 1 is maximal.

ξs: azimuthal peak brightness P.A. measured east of north

x0: location of the ring center horizontally

y0: location of the ring center vertically

Notes

This is a convienence constructor that uses RingTemplate under the hood. To create this function your self do

julia modify(SlashedGaussianRing(σ/r0, s), Rotate(ξs-ξτ), Stretch(r0*sqrt(1-τ), r0/sqrt(1-τ)), Rotate(ξτ), Shift(x0, y0) )

VIDA.GaussianRing — Method

GaussianRing(r0, σ, x0, y0)

Implements the gaussian ring template.

Notes

This is a convienence constructor that uses RingTemplate under the hood. To create this function your self do

modify(GaussianRing(σ/r0), Stretch(r0), Shift(x0, y0))

Arguments

r0: radius of the ring

σ : standard deviation of the Gaussian ring

x0: location of the ring center horizontally

y0: location of the ring center vertically

VIDA.GaussianRing — Method

GaussianRing(σ)

Implements the gaussian ring template.

Notes

This is a convienence constructor that uses RingTemplate under the hood. To create this function your self do

RingTemplate(RadialGaussian(σ), AzimuthalUniform())

Arguments

σ : standard deviation of the Gaussian ring

VIDA.SlashedGaussianRing — Method

SlashedGaussianRing(r0, σ, s, ξ, x0, y0)

Implements the slashed gaussian ring template, that uses a cosine to implement the slash.

Arguments

r0: radius of the ring

σ : standard deviation of the Gaussian ring

s : Slash amplitude. 0 means no slash, and 1 is maximal.

ξs: azimuthal peak brightness P.A. measured east of north

x0: location of the ring center horizontally

y0: location of the ring center vertically

Notes

This is a convienence constructor that uses RingTemplate under the hood. To create this function your self do

modify(SlashedGaussianRing(σ/r0, s), Stretch(r0), Rotate(ξ), Shift(x0, y0))

VIDA.SlashedGaussianRing — Method

SlashedGaussianRing(σ, s)

Implements the slashed gaussian ring template, that uses a cosine to implement the slash and has a brightness position angle of 0 degrees east of north.

Notes

This is a convienence constructor that uses RingTemplate under the hood. To create this function your self do

Arguments

σ : standard deviation of the Gaussian ring

s : Slash amplitude. 0 means no slash, and 1 is maximal.

RingTemplate(RadialGaussian(σ), AzimuthalCosine((s,), (zero(s),)))

VIDA.blur — Method

blur(img, fwhm)

Blurs the img with a gaussian kernel with fwhm in μas. If fwhm is a scalar then the kernel is assumed to be symmetric, otherwise you the first entry is the fwhm in the EW direction and second the NS direction.

Returns the blurred image.

VIDA.blur — Method

blur(mov, fwhm)

Blurs the mov with a gaussian kernel with fwhm in μas. If fwhm is a scalar then the kernel is assumed to be symmetric, otherwise you the first entry is the fwhm in the EW direction and second the NS direction.

Returns the blurred movie.

VIDA.clipimage — Function

clipimage(clip, image)
clipimage(clip, image, mode)

Clips the image im according to the value clip. There are two modes for image clipping: - :relative which zeros the pixels whose intensity are below clip relative to the max. - :absolute which zeros the pixels whose intensity is below clip in Jy/pixel

VIDA.divergence — Method

divergence(d, m)

Computes the divergence of d with respect to the template m. This measure how similar the divergence with respect to the image stored in d is compared to the template.

Arguments

d: An divergence or a subtype of VIDA.AbstractDivergence

m: A template which ComradeBase.AbstractModel or more commonly a VIDA.AbstractImageTemplate

VIDA.divergence_point — Function

divergence_point(div::AbstractDivergence, p, q)

Evaluate the divergence div at a single pixel with value p and q for image and template. The total divergence

VIDA.get_frames — Method

get_frames(mov)

Gets all the frames of the movie object mov. This returns a array of EHTImage objects.

VIDA.get_image — Method

get_image(mov, t; keeptime)

Gets the frame of the movie object mov at the time t. This returns an EHTImage object at the requested time. The returned object is found by linear interpolation.

VIDA.get_times — Method

get_times(mov)

Returns the times that the movie object mov was created at. This does not have to be uniform in time.

VIDA.join_frames — Method

join_frames(times, images)

Joins an array of IntensityMap at specified times to form an VIDAMovie object.

Inputs

times: An array of times that the image was created at

images: An array of EHTImage objects

Outputs

VIDAMovie object

VIDA.load_hdf5 — Method

load_hdf5(filename; style)

Loads an hdf5 file where filename should be a HDF5 file.

Details

This reads in a hdf5 file and outputs and EHTMovie object.

Notes

Currently this only works with movies created by ehtim. SMILI uses a different format, as does Illinois, and every other group...

VIDA.load_im_h5 — Method

load_im_h5(fname)

where fname should be a hdf5 image file generated using illinois hdf5 process

Details

The function returns an EHTImage object that contains the relevant image and parameters extracted from the fits file. It also ensures that we are astronomers and that the image using sky-left coordinates.

VIDA.normalize_div — Function

normalize_div(div, val)

Returns the proper divergence whose values have been normalized so that the value is always postive semi-definite and is only zero then the template and the image are equal.

VIDA.save_hdf5 — Method

save_hdf5(filename, mov; style)

Saves and hdf5 file where filename is the write out location. Currently style only works with ehtim, namely we save HDF5 files that only work with ehtim.

VIDA.triptic — Function

triptic(img::SpatialIntensityMap, template)

Plots a triptic where the left panel is the img middle the template and the right a two cross-sections of the image and template

VIDA.vida — Method

vida(
    prob,
    optimizer;
    rng,
    init_params,
    unit_cube,
    kwargs...
)

Runs the VIDA algorithm to find the optimal template defined by the problem prob. The optimization will be run using the Optimization.jl optimizer. You can optionally pass a random number generator used to specify the initial points or can explicitly pass the initial location of the optimization using init_params. By default vida first transforms the parameter space to the unit hypercube. If you do not wish to do this set unit_cube = false.

The remaining kwargs... are forwarded to the Optimization.solve function.

Arguments

prob: Defines the problem you wish to solve.

optimizer: Specifies the optimizer you want to use. Any optimizer from Optimization.jl works.

Optional Keyword arguments

rng: Specifies the random number generator you want to use to select the initial points. Note that is not forwarded to the optimizers since not all can use a specific rng.

init_params: Specify the initial point of the optimization routine. The default nothing will randomly draw a starting point from the parameter space

unit_cube: If true the parameters are first transformed to the unit hypercube. If false they are transformed to ℝⁿ using TransformVariables

kwargs...: Additional options to be passed to the solve function from Optimization.jl

VLBISkyModels.regrid — Method

regrid(mov::VIDAMovie, npix, xlim, ylim)

Inputs

mov::VIDAMovie : Movie you want to regrid

npix : Number of pixels in x and y direction

xlim : Tuple with the limits of the image in the RA

ylim : Tuple with the limits of the image in DEC